# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Currency service API in Go using Gin framework",
        "title": "Currency Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "paths": {
        "/currency/date1/code": {
            "get": {
                "description": "Return list of currencies.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "currency1"
                ],
                "summary": "Get currency from DB.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set date for currency",
                        "name": "date1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Set code for currency",
                        "name": "code",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "obejct"
                        }
                    }
                }
            }
        },
        "/currencys/date1": {
            "get": {
                "description": "Return list of currencies.",
                "produces": [
                    "application/xml"
                ],
                "tags": [
                    "currency1"
                ],
                "summary": "Get All Currencirs from API.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set date for currency",
                        "name": "date1",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "obejct"
                        }
                    }
                }
            }
        }
    }
}